Rename index files
