+++ src/server/common/helpers/metrics.js
@@ -4,7 +4,7 @@ import {
   StorageResolution
 } from 'aws-embedded-metrics'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { createLogger } from '~/src/server/common/helpers/logging/logger.js'
 
 /**
