+++ src/server/common/helpers/metrics.test.js
@@ -1,6 +1,6 @@
 import { StorageResolution, Unit } from 'aws-embedded-metrics'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { metricsCounter } from '~/src/server/common/helpers/metrics.js'
 
 const mockPutMetric = jest.fn()
