+++ src/server/common/helpers/secure-context/secure-context.js
@@ -1,5 +1,5 @@
 import tls from 'node:tls'
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { getTrustStoreCerts } from '~/src/server/common/helpers/secure-context/get-trust-store-certs.js'
 
 /**
