+++ src/server/common/helpers/session-cache/cache-engine.js
@@ -2,7 +2,7 @@ import { buildRedisClient } from '~/src/server/common/helpers/redis-client.js'
 import { Engine as CatboxRedis } from '@hapi/catbox-redis'
 import { Engine as CatboxMemory } from '@hapi/catbox-memory'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 import { createLogger } from '~/src/server/common/helpers/logging/logger.js'
 
 /**
