+++ src/server/common/helpers/session-cache/session-cache.js
@@ -1,6 +1,6 @@
 import yar from '@hapi/yar'
 
-import { config } from '~/src/config/index.js'
+import { config } from '~/src/config/config.js'
 
 const sessionConfig = config.get('session')
