+++ src/server/common/test-helpers/component-helpers.js
@@ -3,7 +3,7 @@ import path from 'path'
 import nunjucks from 'nunjucks'
 import { load } from 'cheerio'
 import { camelCase } from 'lodash'
-import * as filters from '~/src/config/nunjucks/filters/index.js'
+import * as filters from '~/src/config/nunjucks/filters/filters.js'
 import * as globals from '~/src/config/nunjucks/globals.js'
 
 const dirname = path.dirname(fileURLToPath(import.meta.url))
