+++ src/server/index.js
@@ -1,8 +1,8 @@
 import path from 'path'
 import hapi from '@hapi/hapi'
 
-import { config } from '~/src/config/index.js'
-import { nunjucksConfig } from '~/src/config/nunjucks/index.js'
+import { config } from '~/src/config/config.js'
+import { nunjucksConfig } from '~/src/config/nunjucks/nunjucks.js'
 import { router } from './router.js'
 import { requestLogger } from '~/src/server/common/helpers/logging/request-logger.js'
 import { catchAll } from '~/src/server/common/helpers/errors.js'
